requirements.txt
faker
python-telegram-bot
